+++ ZigBeeNet/ZDO/Command/IeeeAddressResponse.cs
@@ -24,12 +24,12 @@ namespace ZigBeeNet.ZDO.Command
         /**
          * NWKAddrRemoteDev command message field.
          */
-        public int NwkAddrRemoteDev { get; set; }
+        public ushort NwkAddrRemoteDev { get; set; }
 
         /**
          * StartIndex command message field.
          */
-        public int StartIndex { get; set; }
+        public byte StartIndex { get; set; }
 
         /**
          * NWKAddrAssocDevList command message field.
@@ -76,15 +76,15 @@ namespace ZigBeeNet.ZDO.Command
             }
 
             IeeeAddrRemoteDev = (IeeeAddress)deserializer.Deserialize(ZclDataType.Get(DataType.IEEE_ADDRESS));
-            NwkAddrRemoteDev = (int)deserializer.Deserialize(ZclDataType.Get(DataType.NWK_ADDRESS));
+            NwkAddrRemoteDev = (ushort)deserializer.Deserialize(ZclDataType.Get(DataType.NWK_ADDRESS));
 
             if (deserializer.IsEndOfStream)
             {
                 return;
             }
 
-            int? numAssocDev = (int?)deserializer.Deserialize(ZclDataType.Get(DataType.UNSIGNED_8_BIT_INTEGER));
-            StartIndex = (int)deserializer.Deserialize(ZclDataType.Get(DataType.UNSIGNED_8_BIT_INTEGER));
+            byte? numAssocDev = (byte?)deserializer.Deserialize(ZclDataType.Get(DataType.UNSIGNED_8_BIT_INTEGER));
+            StartIndex = (byte)deserializer.Deserialize(ZclDataType.Get(DataType.UNSIGNED_8_BIT_INTEGER));
 
             if (numAssocDev != null)
             {
