+++ ZigBeeNet/ZDO/Command/NetworkAddressRequest.cs
@@ -30,12 +30,12 @@ namespace ZigBeeNet.ZDO.Command
          * 0x01 – Extended response
          * 0x02-0xFF – reserved
          */
-        public int RequestType { get; set; }
+        public byte RequestType { get; set; }
 
         /**
          * StartIndex command message field.
          */
-        public int StartIndex { get; set; }
+        public byte StartIndex { get; set; }
 
         /**
          * Default constructor.
@@ -59,8 +59,8 @@ namespace ZigBeeNet.ZDO.Command
             base.Deserialize(deserializer);
 
             IeeeAddr = (IeeeAddress)deserializer.Deserialize(ZclDataType.Get(DataType.IEEE_ADDRESS));
-            RequestType = (int)deserializer.Deserialize(ZclDataType.Get(DataType.UNSIGNED_8_BIT_INTEGER));
-            StartIndex = (int)deserializer.Deserialize(ZclDataType.Get(DataType.UNSIGNED_8_BIT_INTEGER));
+            RequestType = (byte)deserializer.Deserialize(ZclDataType.Get(DataType.UNSIGNED_8_BIT_INTEGER));
+            StartIndex = (byte)deserializer.Deserialize(ZclDataType.Get(DataType.UNSIGNED_8_BIT_INTEGER));
         }
 
         public bool IsTransactionMatch(ZigBeeCommand request, ZigBeeCommand response)
